requirements.txt
python-dotenv==1.0.1
google-generativeai==0.7.2
streamlit==1.38.0
PyPDF2==3.0.1
joblib==1.4.2